refactor(scoring): reorganize scoring classes into Common namespace and update related tests
