+++ lib/DartsScorer.Main/Scoring/Common/Multiplier.cs
@@ -1,4 +1,4 @@
-namespace DartsScorer.Main.Scoring;
+namespace DartsScorer.Main.Scoring.Common;
 
 public enum Multiplier
 {
