+++ lib/DartsScorer.Main/Scoring/ThrowScore.cs
@@ -1,3 +1,5 @@
+using DartsScorer.Main.Scoring.Common;
+
 namespace DartsScorer.Main.Scoring;
 
 public class ThrowScore
