+++ lib/tests/DartsScorer.Tests/Scoring/LegTests.cs
@@ -1,7 +1,8 @@
 using DartsScorer.Main.Scoring.x01;
+using DartsScorer.Main.Scoring.Common;
 using DartsScorer.Main.Scoring;
 
-namespace DartsScorer.Tests;
+namespace DartsScorer.Tests.Scoring;
 
 public class LegTests
 {
