+++ lib/tests/DartsScorer.Tests/Scoring/ThrowScoreTests.cs
@@ -1,6 +1,7 @@
+using DartsScorer.Main.Scoring.Common;
 using DartsScorer.Main.Scoring;
 
-namespace DartsScorer.Tests;
+namespace DartsScorer.Tests.Scoring;
 
 public class ThrowScoreTests
 {
